+++ ftimage.h
@@ -331,7 +331,7 @@ FT_BEGIN_HEADER
   /*                  it indicates a third-order Bezier arc control point; */
   /*                  and a second-order control point if unset.           */
   /*                                                                       */
-  /*    contours   :: An array of `n_contours' shorts, giving the end      */
+  /*    contours   :: An array of `n_contours' ints, giving the end        */
   /*                  point of each contour within the outline.  For       */
   /*                  example, the first contour is defined by the points  */
   /*                  `0' to `contours[0]', the second one is defined by   */
@@ -343,12 +343,12 @@ FT_BEGIN_HEADER
   /*                                                                       */
   typedef struct  FT_Outline_
   {
-    short       n_contours;      /* number of contours in glyph        */
-    short       n_points;        /* number of points in the glyph      */
+    int         n_contours;      /* number of contours in glyph        */
+    int         n_points;        /* number of points in the glyph      */
 
     FT_Vector*  points;          /* the outline's points               */
     char*       tags;            /* the points flags                   */
-    short*      contours;        /* the contour end points             */
+    int*        contours;        /* the contour end points             */
 
     int         flags;           /* outline masks                      */
